Integrate controller-gen into go.mod, using go generate
Puts controller-gen into version control of go.mod and renovate instead of Makefile
